Moved function to library
